discovery: update ServerData to version 5 with varint encoding
Rewrite ServerData binary format to match the latest Minecraft: Bedrock
Edition wire protocol (version 5). Key changes:

- Use varint encoding for strings, GameType, TransportLayer, ConnectionType
- Change GameType, TransportLayer, ConnectionType types from uint8 to int32
- Add AcceptsOnlineAuth and AcceptsSelfSignedAuth fields
- Add GameType and TransportLayer named constants
- Remove maxServerDataNameLength constraint (varint length prefix)
- Update tests and examples to use new types and constants

+++ discovery/listener.go
@@ -372,9 +372,9 @@ func (l *Listener) PongData(b []byte) {
 	l.ServerData(d)
 }
 
-// parsePongGameType converts the game type string from the pong data to its uint8 representation.
+// parsePongGameType converts the game type string from the pong data to its int32 representation.
 // Returns 0 and false if the game type is not valid.
-func parsePongGameType(v string) (uint8, bool) {
+func parsePongGameType(v string) (int32, bool) {
 	switch strings.ToLower(strings.TrimSpace(v)) {
 	case "survival":
 		return 0, true
